refactor: 타입 파일 용도에 맞게 재분류
main.d.ts   -> habit.d.ts
props.d.ts  -> componentProps.d.ts
(new file)  -> eventHandlers.d.ts
(new file)  -> context.d.ts
