+++ src/interface/habit.d.ts
@@ -1,0 +1,8 @@
+export interface Habit {
+  id: number;
+  name: string;
+  description: string;
+  days: Day[]; // 특정 요일에만 진행하는 경우
+}
+
+export type Day = '월' | '화' | '수' | '목' | '금' | '토' | '일';
